Fixed crash on fullscreen switch, fixed fullscreen at startup not updating resolution, fixed some packets (not all yet)

+++ api/src/main/java/io/xol/chunkstories/api/input/Mouse.java
@@ -27,7 +27,7 @@ public interface Mouse {
 	
 	public float getCursorY();
 	
-	public void setMouseCursorLocation(float x, float y);
+	public void setMouseCursorLocation(double x, double y);
 	
 	public boolean isGrabbed();
